Update relating to significant changes/improvements by owner
When I wrote this review back in 2017, I had some issues with Vapor95, but they've come a loooong way since then, and are now a larger company and have made many reforms to get to where they are. I think that nowadays this review probably doesn't need to exist because of the improvements that have been made, and the legitimate consumer reviews that exist out there. I'm keeping it for now though, so the history is chronicled. Props to Vapor95 and the owner for coming this far and improving so much.